+++ AttackSim.h
@@ -23,7 +23,7 @@ class AttackSim {
 public:
     explicit AttackSim();
     explicit AttackSim(int attFunction, double percentAtt, bool verbose = false);
-    explicit AttackSim(Topology& T, vector<double>& hyperParameters, string fName, string params, int attFunction, double percentAtt, bool verbose = false);
+    explicit AttackSim(Topology& T, vector<int>& hyperParameters, string fName, string params, int attFunction, double percentAtt, bool verbose = false);
     bool split(string input, char del, string &c);
     void performTowerAttack(vector<vector<int>>& connections, vector<bool>& attackedTowers);
     void performDataAttack(vector<vector<double>>& data, vector<bool>& attacktedTowers, int maxOut = 10000, int upper = 20, int lower = 5);
